[REF] sale_coupon_multi_use: LIC/naming/misc/lang
