+++ sale_coupon_multi_use/models/__init__.py
@@ -1,2 +1,3 @@
 from . import sale_coupon_program
 from . import sale_coupon
+from . import sale_coupon_consumption_line
