+++ sale_coupon_multi_use/wizards/sale_coupon_apply_code.py
@@ -1,3 +1,5 @@
+# Copyright 2020 Camptocamp SA
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 from odoo import models
